feat: Organize test files into unit and integration directories
Moved test files from the `tests/` directory into `tests/unit/` and `tests/integration/` subdirectories for better organization and clarity.

- `tests/extractor_unit_tests.rs` -> `tests/unit/extractor_unit_tests.rs`
- `tests/game_unit_tests.rs` -> `tests/unit/game_unit_tests.rs`
- `tests/ascii_art_coverage_tests.rs` -> `tests/integration/ascii_art_coverage_tests.rs`
- `tests/comment_processing_tests.rs` -> `tests/integration/comment_processing_tests.rs`
- `tests/indent_treesitter_tests.rs` -> `tests/integration/indent_treesitter_tests.rs`
- `tests/missing_ascii_art_test.rs` -> `tests/integration/missing_ascii_art_test.rs`

Updated `tests/mod.rs`, `tests/unit/mod.rs`, and `tests/integration/mod.rs` to reflect the new module structure.
