+++ tests/integration/ascii_art_coverage_tests.rs
