+++ tests/integration/missing_ascii_art_test.rs
