+++ tests/integration/mod.rs
@@ -1,11 +1,4 @@
-pub mod languages;
-pub mod multibyte_typing_core;
-
-use gittype::extractor::ExtractionOptions;
-
-pub fn test_extraction_options() -> ExtractionOptions {
-    let mut options = ExtractionOptions::default();
-    // Remove tmp/** pattern for tests since we're using temp directories
-    options.exclude_patterns.retain(|p| p != "**/tmp/**");
-    options
-}
+pub mod ascii_art_coverage_tests;
+pub mod comment_processing_tests;
+pub mod indent_treesitter_tests;
+pub mod missing_ascii_art_test;
