+++ tests/mod.rs
@@ -1,1 +1,2 @@
 pub mod integration;
+pub mod unit;
