+++ tests/unit/mod.rs
@@ -1,0 +1,2 @@
+pub mod extractor_unit_tests;
+pub mod game_unit_tests;
